+++ src/background/context-menus.js
@@ -47,20 +47,18 @@ import {
 	SETTINGS_KEY,
 	TOAST_WARNING,
 	USER_LANGUAGE,
-} from "/constants.js";
-import { openSettingsPage } from "/functions.js";
-import Tab from "/tab.js";
-import ensureTranslatorAvailability from "/translator.js";
+} from "../core/constants.js";
+import { openSettingsPage } from "../core/functions.js";
+import Tab from "../core/tab.js";
+import ensureTranslatorAvailability from "../core/translator.js";
 import {
 	bg_getCurrentBrowserTab,
 	bg_notify,
 	checkLaunchExport,
 } from "./utils.js";
-import {
-	bg_getCommandLinks,
-	bg_getSalesforceLanguage,
-	bg_getSettings,
-} from "./background.js";
+import { bg_getSettings } from "./storage.js";
+import { bg_getCommandLinks } from "./commands.js";
+import { bg_getSalesforceLanguage } from "./salesforce-language.js";
 
 let areMenuItemsVisible = false;