+++ src/entrypoints/options.js
@@ -1,0 +1,4 @@
+"use strict";
+import "../settings/options.js";
+import "../components/help/help.js";
+import "../components/review-sponsor/review-sponsor.js";