+++ docs/CHANGELOG.md
@@ -14,6 +14,11 @@ Please remember that this project is entirely funded by **the support of its use
 ## 🧪 Tests
 -->
 
+# v2.2.5
+
+## 🧪 Tests
+1. Added Logging Auditor & tests
+
 # v2.2.4
 
 ## 🛠 Changed